Move all combination components into their dedicated folder
