+++ admin-dev/themes/new-theme/js/pages/product/combination/combinations-grid-renderer.ts
